doModfy, doDelete 권한체크

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -81,6 +81,7 @@ public class UsrArticleController {
 	public ResultData<Integer> doModify(HttpSession httpSession, int id, String title, String body) {
 
 		boolean isLogined = false;
+		int loginedMemberId = 0;
 
 		if (httpSession.getAttribute("loginedMemberId") != null) {
 			isLogined = true;
@@ -94,6 +95,10 @@ public class UsrArticleController {
 
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 글은 존재하지 않습니다", id), id);
+		}
+		
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", Ut.f("%d번 글에 대한 권한이 없습니다", id), id);
 		}
 
 		articleService.modifyArticle(id, title, body);
@@ -106,9 +111,11 @@ public class UsrArticleController {
 	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
 
 		boolean isLogined = false;
+		int loginedMemberId = 0;
 
 		if (httpSession.getAttribute("loginedMemberId") != null) {
 			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
 		}
 
 		if (isLogined == false) {
@@ -119,6 +126,10 @@ public class UsrArticleController {
 
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 글은 존재하지 않습니다", id), id);
+		}
+		
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", Ut.f("%d번 글에 대한 권한이 없습니다", id), id);
 		}
 
 		articleService.deleteArticle(id);
